Rename to nixops_packet
