+++ nixops_packet/nix/lib.nix
