+++ nixops_packet/plugin.py
@@ -1,7 +1,7 @@
 import os.path
 import nixops.plugins
 import nixops.script_defs
-import nixopspacket.parser
+import nixops_packet.parser
 
 
 @nixops.plugins.hookimpl
@@ -12,10 +12,10 @@ def nixexprs():
 @nixops.plugins.hookimpl
 def load():
     return [
-        "nixopspacket.resources",
-        "nixopspacket.backends.device",
-        "nixopspacket.resources.keypair",
-        "nixopspacket.parser",
+        "nixops_packet.resources",
+        "nixops_packet.backends.device",
+        "nixops_packet.resources.keypair",
+        "nixops_packet.parser",
     ]
 
 
@@ -30,7 +30,7 @@ def parser(parser, subparsers):
         "sos-console",
         help="connect to the machine's sos console",
     )
-    plugin_command.set_defaults(op=nixopspacket.parser.parse_defs.op_sos_console)
+    plugin_command.set_defaults(op=nixops_packet.parser.parse_defs.op_sos_console)
     plugin_command.add_argument(
         "machine", metavar="MACHINE", help="identifier of the machine"
     )
@@ -41,7 +41,7 @@ def parser(parser, subparsers):
         "update-provision",
         help="pull an updated system.nix from a provisioned system",
     )
-    plugin_command.set_defaults(op=nixopspacket.parser.parse_defs.op_update_provision)
+    plugin_command.set_defaults(op=nixops_packet.parser.parse_defs.op_update_provision)
     plugin_command.add_argument(
         "machine", metavar="MACHINE", help="identifier of the machine"
     )
@@ -52,19 +52,19 @@ def parser(parser, subparsers):
         "reinstall",
         help="deprovision, erase, and reinstall an already provisioned system",
     )
-    plugin_command.set_defaults(op=nixopspacket.parser.parse_defs.op_reinstall)
+    plugin_command.set_defaults(op=nixops_packet.parser.parse_defs.op_reinstall)
     plugin_command.add_argument(
         "machine", metavar="MACHINE", help="identifier of the machine"
     )
     nixops.script_defs.add_common_deployment_options(plugin_command)
 
     #    plugin_command = plugin_cmd_subparsers.add_parser('foo', help='execute command "foo"')
-    #    plugin_command.set_defaults(op=nixopspacket.parser.parse_defs.op_foo)
+    #    plugin_command.set_defaults(op=nixops_packet.parser.parse_defs.op_foo)
     #    plugin_command.add_argument('--verbose', '-v', action='store_true', help='Provide extra foo information')
     #    plugin_command.add_argument('--debug', action='store_true', help='enable debug output')
 
     #    plugin_command = plugin_cmd_subparsers.add_parser('bar', help='execute command "bar"')
-    #    plugin_command.set_defaults(op=nixopspacket.parser.parse_defs.op_bar)
+    #    plugin_command.set_defaults(op=nixops_packet.parser.parse_defs.op_bar)
     #    plugin_command.add_argument('--verbose', '-v', action='store_true', help='Provide extra bar information')
     #    plugin_command.add_argument('--debug', action='store_true', help='enable debug output')
     return
