+++ nixops_packet/resources/__init__.py
