+++ nixops_packet/utils.py
