+++ pyproject.toml
@@ -1,11 +1,11 @@
 
 [tool.poetry]
-name = "nixopspacket"
+name = "nixops_packet"
 version = "1.0"
 description = "NixOps plugin for Packet.com"
 authors = ["Graham Christensen <graham@grahamc.com>"]
 license = "LGPL-3.0-only"
-include = [ "nixopspacket/nix/*.nix" ]
+include = [ "nixops_packet/nix/*.nix" ]
 
 [tool.poetry.dependencies]
 python = "^3.7"
@@ -19,7 +19,7 @@ mypy = "^0.770"
 black = "^19.10b0"
 
 [tool.poetry.plugins."nixops"]
-packet = "nixopspacket.plugin"
+packet = "nixops_packet.plugin"
 
 [build-system]
 requires = ["poetry>=0.12"]
